+++ src/chainIds.ts
@@ -1,0 +1,30 @@
+export enum ChainsID {
+  Unset = 0,
+  Solana = 1,
+  Ethereum = 2,
+  Terra = 3,
+  BSC = 4,
+  Polygon = 5,
+  Avalanche = 6,
+  Oasis = 7,
+  Algorand = 8,
+  Aurora = 9,
+  Fantom = 10,
+  Karura = 11,
+  Acala = 12,
+  Klaytn = 13,
+  Celo = 14,
+  Near = 15,
+  Moonbeam = 16,
+  Neon = 17,
+  Terra2 = 18,
+  Injective = 19,
+  Sui = 21,
+  Aptos = 22,
+  Arbitrum = 23,
+  Optimism = 24,
+  PythNet = 26,
+  Xpla = 28,
+  Btc = 29,
+  Wormchain = 3104,
+}
